+++ tsinfer/formats.py
@@ -28,7 +28,6 @@ import os.path
 import queue
 import sys
 import threading
-import uuid
 import warnings
 
 import attr
@@ -385,7 +384,6 @@ class DataContainer:
             self.data = zarr.open_group(store=store, mode="w")
         self.data.attrs[FORMAT_NAME_KEY] = self.FORMAT_NAME
         self.data.attrs[FORMAT_VERSION_KEY] = self.FORMAT_VERSION
-        self.data.attrs["uuid"] = str(uuid.uuid4())
 
         chunks = self._chunk_size
         provenances_group = self.data.create_group("provenances")
@@ -486,8 +484,7 @@ class DataContainer:
         """
         Returns a copy of this DataContainer opened in 'edit' mode. If path
         is specified, this must not be equal to the path of the current
-        data container. The new container will have a different UUID to the
-        current.
+        data container.
         """
         if self._mode != self.READ_MODE:
             raise ValueError("Cannot copy unless in read mode.")
@@ -516,8 +513,6 @@ class DataContainer:
             store = other._new_lmdb_store(max_file_size)
             zarr.copy_store(self.data.store, store)
             other.data = zarr.group(store)
-        # Set a new UUID
-        other.data.attrs["uuid"] = str(uuid.uuid4())
         other.data.attrs[FINALISED_KEY] = False
         other._mode = self.EDIT_MODE
         return other
@@ -664,10 +659,6 @@ class DataContainer:
         return ret
 
     @property
-    def uuid(self):
-        return str(self.data.attrs["uuid"])
-
-    @property
     def num_provenances(self):
         return self.provenances_timestamp.shape[0]
 
@@ -693,7 +684,7 @@ class DataContainer:
     def __eq__(self, other):
         ret = NotImplemented
         if isinstance(other, type(self)):
-            ret = self.uuid == other.uuid and self.data_equal(other)
+            ret = self.data_equal(other)
         return ret
 
     def __str__(self):
@@ -703,7 +694,6 @@ class DataContainer:
             ("format_name", self.format_name),
             ("format_version", self.format_version),
             ("finalised", self.finalised),
-            ("uuid", self.uuid),
             ("num_provenances", self.num_provenances),
             ("provenances/timestamp", zarr_summary(self.provenances_timestamp)),
             ("provenances/record", zarr_summary(self.provenances_record)),
@@ -1320,10 +1310,10 @@ class SampleData(DataContainer):
         """
         Returns True if all the data attributes of this input file and the
         specified input file are equal. This compares every attribute except
-        the UUID and provenance.
+        the provenance.
 
         To compare two :class:`SampleData` instances for exact equality of
-        all data including UUIDs and provenance data, use ``s1 == s2``.
+        all data including provenance data, use ``s1 == s2``.
 
         :param SampleData other: The other :class:`SampleData` instance to
             compare with.
@@ -2100,7 +2090,7 @@ class SampleData(DataContainer):
                 genos = geno_map[genos]
             yield Variant(site=site, alleles=alleles, genotypes=genos)
 
-    def __all_haplotypes(self, sites=None, recode_ancestral=None):
+    def _all_haplotypes(self, sites=None, recode_ancestral=None):
         # We iterate over chunks vertically here, and it's not worth complicating
         # the chunk iterator to handle this.
         if recode_ancestral is None:
@@ -2156,7 +2146,7 @@ class SampleData(DataContainer):
                 raise ValueError("Sample index too large.")
         j = 0
 
-        for index, a in self.__all_haplotypes(sites, recode_ancestral):
+        for index, a in self._all_haplotypes(sites, recode_ancestral):
             if j == len(samples):
                 break
             if index == samples[j]:
@@ -2225,6 +2215,257 @@ class SampleData(DataContainer):
             yield Population(j, metadata=metadata)
 
 
+class SgkitSampleData(SampleData):
+
+    FORMAT_NAME = "tsinfer-sgkit-sample-data"
+    FORMAT_VERSION = (0, 1)
+
+    def __init__(self, path):
+        self.path = path
+        self.data = zarr.open(path, mode="r")
+        genotypes_arr = self.data["call_genotype"]
+        self._num_sites, self._num_individuals, self.ploidy = genotypes_arr.shape
+        self._num_samples = self._num_individuals * self.ploidy
+
+        assert self.ploidy == self.data["call_genotype"].chunks[2]
+
+    def __metadata_schema_getter(self, zarr_group):
+        try:
+            return self.data[zarr_group].attrs["metadata_schema"]
+        except KeyError:
+            return {"codec": "json"}
+
+    @property
+    def format_name(self):
+        return self.FORMAT_NAME
+
+    @property
+    def format_version(self):
+        return self.FORMAT_VERSION
+
+    @property
+    def finalised(self):
+        return True
+
+    @property
+    def sequence_length(self):
+        try:
+            return self.data.attrs["sequence_length"]
+        except KeyError:
+            return int(np.max(self.data["variant_position"])) + 1
+
+    @property
+    def num_sites(self):
+        return self._num_sites
+
+    @property
+    def sites_metadata_schema(self):
+        return self.__metadata_schema_getter("sites")
+
+    @property
+    def sites_metadata(self):
+        try:
+            return self.data["sites/metadata"]
+        except KeyError:
+            return zarr.array(
+                [{}] * self.num_individuals, object_codec=numcodecs.JSON()
+            )
+
+    @property
+    def sites_time(self):
+        try:
+            return self.data["sites/time"]
+        except KeyError:
+            return np.full(self.data["variant_position"].shape, tskit.UNKNOWN_TIME)
+
+    @property
+    def sites_position(self):
+        return self.data["variant_position"]
+
+    @property
+    def sites_alleles(self):
+        return self.data["variant_allele"]
+
+    @property
+    def sites_ancestral_allele(self):
+        try:
+            return self.data["variant_ancestral_allele_index"]
+        except KeyError:
+            # Maintains backwards compatibility: in previous tsinfer versions the
+            # ancestral allele was always the zeroth element in the alleles list
+            return np.zeros(self.num_sites, dtype=np.int8)
+
+    @property
+    def sites_genotypes(self):
+        gt = self.data["call_genotype"]
+        # This method is only used for test/debug so we retrive and
+        # reshape the entire array.
+        return gt[...].reshape(gt.shape[0], gt.shape[1] * gt.shape[2])
+
+    @property
+    def provenances_timestamp(self):
+        try:
+            return self.data["provenances_timestamp"]
+        except KeyError:
+            return np.array([], dtype=object)
+
+    @property
+    def provenances_record(self):
+        try:
+            return self.data["provenances_record"]
+        except KeyError:
+            return np.array([], dtype=object)
+
+    @property
+    def num_samples(self):
+        return self._num_samples
+
+    @property
+    def samples_individual(self):
+        ret = np.zeros((self.num_samples), dtype=np.int32)
+        for p in range(self.ploidy):
+            ret[p :: self.ploidy] = np.arange(self.num_individuals)
+        return ret
+
+    @property
+    def metadata_schema(self):
+        try:
+            return self.data.attrs["metadata_schema"]
+        except KeyError:
+            None
+
+    @property
+    def metadata(self):
+        try:
+            return self.data.attrs["metadata"]
+        except KeyError:
+            return b""
+
+    @property
+    def populations_metadata(self):
+        try:
+            return self.data["populations/metadata"]
+        except KeyError:
+            return np.array([], dtype=object)
+
+    @property
+    def populations_metadata_schema(self):
+        return self.__metadata_schema_getter("populations")
+
+    @property
+    def num_individuals(self):
+        return self._num_individuals
+
+    @property
+    def individuals_time(self):
+        try:
+            return self.data["individuals/time"]
+        except KeyError:
+            return np.full(self.num_individuals, tskit.UNKNOWN_TIME)
+
+    @property
+    def individuals_metadata_schema(self):
+        return self.__metadata_schema_getter("individuals")
+
+    @property
+    def individuals_metadata(self):
+        try:
+            return self.data["individuals/metadata"]
+        except KeyError:
+            return zarr.array(
+                [{}] * self.num_individuals, object_codec=numcodecs.JSON()
+            )
+
+    @property
+    def individuals_location(self):
+        try:
+            return self.data["individuals/location"]
+        except KeyError:
+            return zarr.array([[]] * self.num_individuals, dtype=float)
+
+    @property
+    def individuals_population(self):
+        try:
+            return self.data["individuals/population"]
+        except KeyError:
+            return np.full((self.num_individuals), tskit.NULL, dtype=np.int32)
+
+    @property
+    def individuals_flags(self):
+        try:
+            return self.data["individuals/population"]
+        except KeyError:
+            return np.full((self.num_individuals), 0, dtype=np.int32)
+
+    def variants(self, sites=None, recode_ancestral=None):
+        """
+        Returns an iterator over the :class:`Variant` objects. This is equivalent to
+        the :meth:`tskit.TreeSequence.variants` iterator. If recode_ancestral is
+        ``True``, the ``.alleles`` attribute of each variant is guaranteed to return
+        the alleles in an order such that the ancestral state is the first item
+        in the list. In this case, ``variant.alleles`` may list the alleles in a
+        different order from the input order as listed in ``variant.site.alleles``,
+        and the values in genotypes array will be recoded so that the ancestral
+        state will have a genotype of 0. If the ancestral state is unknown, the
+        original input order is kept.
+
+        :param array sites: A numpy array of ascending site ids for which to return
+            data. If None (default) return all sites.
+        :param bool recode_ancestral: If True, recode genotypes at sites where the
+            ancestral state is known such that the ancestral state is coded as 0,
+            as described above. Otherwise return genotypes in the input allele encoding.
+            Default: ``None`` treated as ``False``.
+        :return: An iterator over the variants in the sample data file.
+        :rtype: iter(:class:`Variant`)
+        """
+        if recode_ancestral is None:
+            recode_ancestral = False
+        all_genotypes = chunk_iterator(self.data["call_genotype"], indexes=sites)
+        assert MISSING_DATA < 0  # required for geno_map to remap MISSING_DATA
+        for genos, site in zip(all_genotypes, self.sites(ids=sites)):
+            # We have an extra ploidy dimension when coming from sgkit
+            genos = genos.reshape(self.num_samples)
+            aa = site.ancestral_allele
+            alleles = site.alleles
+            if aa != MISSING_DATA and aa > 0 and recode_ancestral:
+                # Need to recode this site
+                alleles = site.reorder_alleles()
+                # re-map the genotypes
+                geno_map = np.arange(len(alleles) - MISSING_DATA, dtype=genos.dtype)
+                geno_map[MISSING_DATA] = MISSING_DATA
+                geno_map[aa] = 0
+                geno_map[0:aa] += 1
+                genos = geno_map[genos]
+            yield Variant(site=site, alleles=alleles, genotypes=genos)
+
+    def _all_haplotypes(self, sites=None, recode_ancestral=None):
+        # We iterate over chunks vertically here, and it's not worth complicating
+        # the chunk iterator to handle this.
+        if recode_ancestral is None:
+            recode_ancestral = False
+        aa_index = self.sites_ancestral_allele[:]
+        # If ancestral allele is missing, keep the order unchanged (aa_index of zero)
+        aa_index[aa_index == MISSING_DATA] = 0
+        gt = self.data["call_genotype"]
+        chunk_size = gt.chunks[1]
+        for j in range(self.num_individuals):
+            if j % chunk_size == 0:
+                chunk = gt[:, j : j + chunk_size, :]
+            indiv_gt = chunk[:, j % chunk_size, :]
+            for k in range(self.ploidy):
+                a = indiv_gt[:, k].T
+                if recode_ancestral:
+                    # Remap the genotypes at all sites, depending on the aa_index
+                    a = np.where(
+                        a == aa_index,
+                        0,
+                        np.where(
+                            np.logical_and(a != MISSING_DATA, a < aa_index), a + 1, a
+                        ),
+                    )
+                yield (j * self.ploidy) + k, a if sites is None else a[sites]
+
+
 @attr.s(order=False, eq=False)
 class Ancestor:
     """
@@ -2291,7 +2532,6 @@ class AncestorData(DataContainer):
         super().__init__(**kwargs)
         sample_data._check_finalised()
         self.sample_data = sample_data
-        self.data.attrs["sample_data_uuid"] = sample_data.uuid
         if self.sample_data.sequence_length == 0:
             raise ValueError("Bad samples file: sequence_length cannot be zero")
         self.data.attrs["sequence_length"] = self.sample_data.sequence_length
@@ -2371,7 +2611,6 @@ class AncestorData(DataContainer):
     def __str__(self):
         values = [
             ("sequence_length", self.sequence_length),
-            ("sample_data_uuid", self.sample_data_uuid),
             ("num_ancestors", self.num_ancestors),
             ("num_sites", self.num_sites),
             ("sites/position", zarr_summary(self.sites_position)),
@@ -2386,12 +2625,10 @@ class AncestorData(DataContainer):
     def data_equal(self, other):
         """
         Returns True if all the data attributes of this input file and the
-        specified input file are equal. This compares every attribute except
-        the UUID.
+        specified input file are equal. This compares every attribute.
         """
         return (
             self.sequence_length == other.sequence_length
-            and self.sample_data_uuid == other.sample_data_uuid
             and self.format_name == other.format_name
             and self.format_version == other.format_version
             and self.num_ancestors == other.num_ancestors
@@ -2412,10 +2649,6 @@ class AncestorData(DataContainer):
         Returns the sequence length.
         """
         return self.data.attrs["sequence_length"]
-
-    @property
-    def sample_data_uuid(self):
-        return self.data.attrs["sample_data_uuid"]
 
     @property
     def num_ancestors(self):
@@ -2531,14 +2764,7 @@ class AncestorData(DataContainer):
             (i.e. breaking the infinite sites assumption), allowing them to possess
             derived alleles at sites where there are no pre-existing mutations in
             older ancestors.
-        :param bool require_same_sample_data: If ``True`` (default) then the
-            the ``sample_data`` parameter must point to the same :class:`.SampleData`
-            instance as that used to generate the current ancestors. If ``False``,
-            this requirement is not enforced, and it is the user's responsibility
-            to ensure that the encoding of alleles in ``sample_data`` matches the
-            encoding in the current :class:`AncestorData` instance (i.e. that in the
-            original :class:`.SampleData` instance on which the current ancestors
-            are based).
+        :param bool require_same_sample_data: **Deprecated** Has no effect.
         :param \\**kwargs: Further arguments passed to the constructor when creating
             the new :class:`AncestorData` instance which will be returned.
 
@@ -2547,11 +2773,6 @@ class AncestorData(DataContainer):
         """
         self._check_finalised()
         sample_data._check_finalised()
-        if require_same_sample_data:
-            if sample_data.uuid != self.sample_data_uuid:
-                raise ValueError(
-                    "sample_data differs from that used to build the initial ancestors"
-                )
         if self.sequence_length != sample_data.sequence_length:
             raise ValueError("sample_data does not have the correct sequence length")
         used_sites = np.isin(sample_data.sites_position[:], self.sites_position[:])
@@ -2646,8 +2867,6 @@ class AncestorData(DataContainer):
             other.clear_provenances()
             for timestamp, record in self.provenances():
                 other.add_provenance(timestamp, record)
-            if sample_data.uuid != self.sample_data_uuid:
-                pass  # TODO: if sample files don't match, we need extra provenance info
             other.record_provenance(command="insert_proxy_samples", **kwargs)
 
         assert other.num_ancestors == self.num_ancestors + len(sample_ids)